Bye InstallersData.crc_installer:
We can compute it very cheaply. We are saving it on close in case of
downgrading to a previous commit. Only used in converters btw - renamed
some 'data' to 'idata' there (converters need revisiting).

This will allow me to simplify adding an archive vs adding a project
(see next commit), plus it simplifies irefresh.