InstallersData.refresh_installers - fix for ghosts in MEI: TTT EEE maybe use shellMove ?
EEE Copy to conflicts - don't use tmp dir ?

It was high time - note how many ugly lines of code this takes with it.
In particular, in Monitor External Installation, we were calling
ShowPanel just to add the project to the installers data. Also,
_get_refreshed was a hack, all this belonged to the data model, plus
the is_project logic was due to crc_installer being updated in irefresh,
fixed in parent commit. createFromData didn't belong to
InstallersProject (staticmethod is more often than not an indication
that the method belongs elsewhere, not always though).

This was triggered by me trying to remove the ShowPanel refresh in
Monitor External Installation but decided to solve this at the root
Note the fixup for ghosts - if autoghosting is on the modInfos refresh
will ghost the newly added espms and the copyTo operation will fail.

The ci_files is from future (moved this here, was after the LowerDict).